+++ .template-lintrc.js
@@ -513,6 +513,7 @@ module.exports = {
       "moduleId": "tests/dummy/app/templates/ember-flexberry-dummy-multi-list",
       "only": [
         "quotes"
+        "eol-last"
       ]
     },
     {
@@ -848,6 +849,18 @@ module.exports = {
     },
     {
       "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/folv-paging",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
+      "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/folv-user-settings",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
+      "moduleId": "tests/dummy/app/templates/components-acceptance-tests/flexberry-objectlistview/ember-flexberry-dummy-multi-list",
       "only": [
         "eol-last"
       ]