+++ tests/dummy/app/models/ember-flexberry-dummy-application-user.js
@@ -1,3 +1,4 @@
+import Ember from 'ember';
 import DS from 'ember-data';
 import EmberValidations from 'ember-validations';
 import EmberFlexberryDataModel from 'ember-flexberry-data/models/model';