+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.7.0-beta.5] - 2022-02-01
+### Added
+* Optional `ember-cp-validations` support.
+
+### Fixed
+* The `flexberry-objectlistview` component:
+    * Using of `perPage` setting at `developerUserSettings`.
+* Validation for deleted records in `flexberry-groupedit` on `edit-form`.
+
 ## [3.7.0-beta.9] - 2022-01-18
 ### Fixed
 * Null content check in `object-list-view`.